# Updated HDR histogram from upsteam after they merged our fix in #10606.