--- a/deps/hdr_histogram/hdr_histogram.c
+++ b/deps/hdr_histogram/hdr_histogram.c
@@ -165,16 +165,6 @@ static int32_t count_leading_zeros_64(int64_t value)
 #endif
 }
 
-static int64_t get_count_at_index_given_bucket_base_idx(const struct hdr_histogram* h, int32_t bucket_base_idx, int32_t sub_bucket_idx)
-{
-    return h->counts[(bucket_base_idx + sub_bucket_idx) - h->sub_bucket_half_count];
-}
-
-static int32_t get_bucket_base_index(const struct hdr_histogram* h, int32_t bucket_index)
-{
-    return (bucket_index + 1) << h->sub_bucket_half_count_magnitude;
-}
-
 static int32_t get_bucket_index(const struct hdr_histogram* h, int64_t value)
 {
     int32_t pow2ceiling = 64 - count_leading_zeros_64(value | h->sub_bucket_mask); /* smallest power of 2 containing value */
@@ -679,32 +669,20 @@ int64_t hdr_min(const struct hdr_histogram* h)
 static int64_t get_value_from_idx_up_to_count(const struct hdr_histogram* h, int64_t count_at_percentile)
 {
     int64_t count_to_idx = 0;
-    int64_t value_from_idx = 0;
-    int32_t sub_bucket_idx = -1;
-    int32_t bucket_idx = 0;
-    int32_t bucket_base_idx = get_bucket_base_index(h, bucket_idx);
 
-    // Overflow check
-    if (count_at_percentile > h->total_count)
+    count_at_percentile = 0 < count_at_percentile ? count_at_percentile : 1;
+    for (int32_t idx = 0; idx < h->counts_len; idx++)
     {
-        count_at_percentile = h->total_count;
-    }
-
-    while (count_to_idx < count_at_percentile)
-    {
-        // increment bucket
-        sub_bucket_idx++;
-        if (sub_bucket_idx >= h->sub_bucket_count)
+        count_to_idx += h->counts[idx];
+        if (count_to_idx >= count_at_percentile)
         {
-            sub_bucket_idx = h->sub_bucket_half_count;
-            bucket_idx++;
-            bucket_base_idx = get_bucket_base_index(h, bucket_idx);
+            return hdr_value_at_index(h, idx);
         }
-        count_to_idx += get_count_at_index_given_bucket_base_idx(h, bucket_base_idx, sub_bucket_idx);
-        value_from_idx = ((int64_t)(sub_bucket_idx)) << (((int64_t)(bucket_idx)) + h->unit_magnitude);
     }
-    return value_from_idx;
+
+    return 0;
 }
+
 
 int64_t hdr_value_at_percentile(const struct hdr_histogram* h, double percentile)
 {