@yoav-steinberg 2022-05-22
> @yoav-steinberg i understand we don't use `git subtree` for this dependency, so i can just squash-merge it, right? however, i think the README.md file in the deps folder might need an update. it currently says:
> 
> ```
> We use a customized version 0.11.5
> 1. Compare all changes under /hdr_histogram directory to version 0.11.5
> 2. Copy updated files from newer version onto files in /hdr_histogram.
> 3. Apply the changes from 1 above to the updated files.
> ```
> 
> but i see this PR uses their master branch. so maybe we have to refer to a SHA now?

Right, I'll update the README.